requirements.txt
requests==2.31.0
markdown==3.5.2
beautifulsoup4==4.12.3
python-dotenv==1.2.2
